+++ www/front_src/src/Resources/Actions/api/index.ts
@@ -1,22 +1,14 @@
-import axios, { AxiosResponse, CancelToken } from 'axios';
 import formatISO from 'date-fns/formatISO';
+import axios, { AxiosResponse, CancelToken } from 'axios';
 import { map, pick } from 'ramda';
 
-import { getData } from '@centreon/ui';
-
 import {
-  buildResourcesEndpoint,
-  serviceCheckEndpoint,
-  hostCheckEndpoint,
-  acknowledgeEndpoint,
   downtimeEndpoint,
+  hostCheckEndpoint,
+  serviceCheckEndpoint,
+  acknowledgeEndpoint,
 } from './endpoint';
-import { ResourceListing, Resource } from '../models';
-
-const listResources = (cancelToken) => (
-  endpointParams,
-): Promise<ResourceListing> =>
-  getData<ResourceListing>(cancelToken)(buildResourcesEndpoint(endpointParams));
+import { Resource } from '../../models';
 
 interface AcknowledgeParams {
   acknowledgeAttachedResources?: boolean;
@@ -119,10 +111,4 @@ const checkResources = ({
   );
 };
 
-export {
-  acknowledgeResources,
-  setDowntimeOnResources,
-  checkResources,
-  listResources,
-  getData,
-};
+export { acknowledgeResources, setDowntimeOnResources, checkResources };
