+++ www/front_src/src/Resources/Listing/index.test.tsx
@@ -59,6 +59,7 @@ const fillEntities = (): Array<Resource> => {
       index % 5 === 0 ? `Entity ${index}` : `Entity ${index}\n Line ${index}`,
     type: index % 4 === 0 ? 'service' : 'host',
     details_endpoint: 'endpoint',
+    timeline_endpoint: 'endpoint',
   }));
 };
